+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file in
 accordance with
 [![keepachangelog 1.0.0](https://img.shields.io/badge/keepachangelog-1.0.0-brightgreen.svg)](http://keepachangelog.com/en/1.0.0/)
 
+## [0.1.71] - 2025-04-10
+
+### Added
+
+- Examples in CLI help/usage
+- Autogenerate CLI docs for website from actual CLI
+
+### Changed
+
+- CLI version command now outputs to stdout (instead of stderr)
+
+### Fixed
+
+- `sudo: opctl command not found` error if `opctl` not on PATH
+
 ## [0.1.70] - 2025-04-02
 
 ### Added